fix: resolve 22 pre-existing mypy errors, fix 2 ruff errors
Mypy: 30 → 8 errors (remaining 8 are pre-existing object-type vertex issues)
Ruff: 2 → 0 errors

- Fix metrics.py return type (dict[str, object])
- Fix spectrum.py parameter annotations (Digraph not object)
- Fix approximation.py return type
- Fix adaptive_beta.py no-any-return
- Fix networkx.py adapter types (Any instead of object)
- Fix closed_form.py int() cast for vertex iteration
- Fix algorithm.py _sample_pivots accept Iterable
- Fix benchmark scripts row types (dict[str, Any])
- Fix benchmark_reachability.py hop_count return type

+++ reachq/research/adaptive_beta.py
@@ -97,4 +97,4 @@ def paper_beta(graph: Digraph, omega: float = 3.0) -> float:
     m = graph.num_edges()
     if m == 0:
         return float("inf")
-    return (n**omega / m) ** (1.0 / (2.0 * omega - 2.0))
+    return float((n**omega / m) ** (1.0 / (2.0 * omega - 2.0)))
